Build open-vm-tools with xmlsec1 by default.
Build open-vm-tools with xmlsec1 by default, rather than with
xml-security-c as is done currently.  This requires modifications
to configure.ac to change the default for use_xmlsec1, plus
modifications to bora-vmsoft/install/Source/tools-pkg.make to
initialize flags for xmlsec1 and xml2 now needed by the OVT build
that is done during tools builds.

Another change that surfaced during this was that the symbol that
configure checks for the xmlsec library was wrong.  It should be
xmlSecCheckVersionExt rather than xmlSecCheckVersion.  The latter
is a header-defined macro but not a library symbol.