open-vm-tools: build with tirpc
Newer distros (Fedora, OpenSuSE Tumblweed) no longer ship rpcgen with glibc,
so we need to build with libtirpc (see
https://bugzilla.redhat.com/show_bug.cgi?id=1531540 and
https://bugzilla.suse.com/show_bug.cgi?id=1089181).
With this change, configure will check for the presence of libtirpc and
if found, sets flags to build with it. configure will also exit with an
error if rpcgen is not found.