FreeBSD: clang compiler related changes to configure.ac
On FreeBSD, the unwind.h header is available in the libunwind
package (/usr/ports/devel/libunwind) which may or may not be installed
on a system.  The libunwind functions to determine the call-chain in a
program are defined only if " __GNU_SOURCE" is defined to enable "all"
GCC extensions.  Avoid the unwind.h header if seen on FreeBSD.

The "clang" compiler will issue warnings if the address of a member
of a packed structure is taken.  Since warnings are errors with -Wall,
add --Wno-address-of-packed-member to CFLAGS when compiling with
"clang".