[tools-plugin] Potential solution for reported desktopEvent crash
On several distros like RHEL, CentOS and Fedora, plugin desktopEvents.so
crashes have been reported.  However, neither development or QE can
reproduce this issue in local environments.

The XFetchName document states that it may generate BadWindow error.  That
means that the value for a Window argument does not name a defined Window.
X FAQ also suggests handling badwindow error after finding it.  Since
the window does not exist when XFetchName is run on it, the name is not
needed.  Suppress this error to avoid a crash.

After grabbing the X server, use XSetErrorHandler to temporarily define
a local error handler to intercept and cleanly handle any BadWindow
errors.  XSetErrorHandler returns the current error handler, so it is
straightforward to reinstate GDK's original handler.