Add the notification to toolsd about the guest log level change.
Consolidated the VIGOR APIs of guestLogEnabled and guestLogLevel, since
it is easier for the customers to use just one API to set both. Use
the "disable" enum value to disable the guest logging instead of
using a separate field and VIGOR API.

Now sending a new tools option to the toolsd whenever the guestLogLevel
changes.