Calculating monitor work area: No bar exists if seamless maximizing
In Unbuntu 18.04, Redhat 7.4, 7.5, the work area size could not be calculated
based on _NET_WM_STRUT or _NET_WM_STRUT_PARTIAL.  These two properties could
not be retrived through the window manger.  In this fix, use _NET_WORKAREA to
get the work area, but this is only available with a single monitor.