Fix a gcc-8 compiler warning in lib/misc/vthreadBase.c
gcc-8 generates a stringop-truncation warning when it's possible
for strncpy to exclude the trailing nul.  The code was fine, we never
touch the last byte in the buffer and it's a static, but explicitly
set a nul at the end of the buffer so gcc sees it.

This is needed for open-vm-tools to build on SUSE Tumbleweed.