Allow only a single instance of vmusr when multiple users are logged into a VM
When a vmusr process gets the "channel conflict" error while attempting
to open the toolbox-dnd channel, a channel reset is triggered.  That
reset results in the channel being restarted and a subsequent conflict
and reset occurs - every second until the channel becomes available.

For *nix guests:
The fix is making use of the repetitive channel resets where the only
RpcIn message received is a "reset" to catch this channel "permanently"
unavailable state.  If other RpcIn messages are received, a channel
is considered to be working and the cumulative error count is cleared..

lib/rpcin/rpcin.c:
- struct RpcIn: Added error status boolean and callback function to
                notify the dependent layer that a channel error has been
                resolved.
- RpcInLoop(): If a non "reset" message is received, clear any channel
               error status.  This will also notify the dependent layer
               that the channel is functioning.
- RpcIn_start(): Added additional argument for new callback; NULL if
                 not needed.

lib/rpcChannel/rpcChannel.c:
- struct rpcChannelInt:
  - Renamed "rpcErrorCount" to "rpcResetErrorCount" since it is actually
    a count of the consecutive channel reset failures and not a count
    of RpcChannel errors.
  - Added counter "rpcFailureCount" for the cumulative channel errors.
  - Added "rpcFailureCb" for optional callback to notify the app of a
    "permanent" channel failure.
- New function RpcChannelClearError() for RpcIn to notify when the
  channel is working; to clear the rpcFailureCount .
- RpcChannel_Setup() - added two arguments for (1) an optional function
                       to be called when there is a channel failure
                       and (2) a failure count threshold.
                       These optional values are stored in the
                       RpcChannel structure being created.
- RpcChannelError(): Added logic to notify the calling app if the error
                     threshold has been reached and notify the app if a
                     callback was provided. A zero threshold signifies
                     the single vmusr limit should not be enforced.
                     (fix disable switch).

services/vmtoolsd/mainLoop.c:
- New function ToolsCore_GetVmusrLimit() to retrieve the channel error
  threshold default or over-ride from tools.conf.

services/vmtoolsd/toolsRpc.c:
- Added ToolsCoreAppChannelFail(): Callback for "permanent" channel
  connection failure.  A warning is logged based on whether another
  "vmtoolsd -n vmusr" is running or not and the process is terminated.
  On Mac OS, the process is terminated with exit(1) as an indication
  to launchd that the vmusr process should not automatically be
  restarted.

The current implementation uses the error callback only for the vmusr
server on Linux (*nix).
The default channel error limit is 5 (approx. 5 second), but is user
configurable in tools.conf.

   [vmusr]
   maxChannelAttempts = n    # where allowed n = 0, 3-15

When "maxChannelAttempts = 0" is used, the restriction to a single
running vmusr process is not enforced.   The existing behavior is
restored with all the accompanying VMX log spew.  This is essentially
a user configurable feature disablement switch.