Implement tools hang detection logic
Create a dedicated detector thread. The thread sits in a loop and
wakes up periodically to decrement an atomic counter. Also schedules
a checkin timer with the main loop to reset the counter periodically.

If the counter ever drops to/below zero, the tools hang is detected,
and a tools hang event is generated. Otherwise, if there was a hang,
but the	counter has now come back up to positive, a tools recovery
event is generated.

In order to properly create hang and recovery event, previous state
needs to be tracked.

In order to properly handle shutdown, we need a condition variable so
that the detector thread can wake up on it while sleeping. This is
because the toolsd calls the thread pool shutdown function which
in turn calls each thread's terminate function and wait for the threads
to quit. Therefore, our terminate function shall wake up the detector
thread and make it quit. Otherwise, the toolsd shutdown shall hang.

Next change shall implement the new RPCI command to send the
hang/recovery event to VMX.