Develop log APIs to fix security holes in the tools log messages.
Security artifacts such as command args on the host should not be logged
in the VMX log files on the host.  This change creates APIs so that
different log messages can be used for host and guest.

Refactored the log plumbing to minimize code duplication when calling
the different implementation of logging to the vmx and logging in guest.

Fixed one instance of security issue in vmbackup, to show how to use
the new APIs.