Clear DnDPluginResetTimer on receiving SIGUSR1.
DnDPluginResetTimer callback handler DnDPluginResetSent() accesses
the RPC channel managed by the main loop. However, mainloop destroys
this channel on receiving SIGUSR1. So, we need to destroy this timer
as well when SIGUSR1 is received.

There are at least 2 ways to do it:
1. Add Linux specific code in vmCopyPasteDnDWrapper.cpp and any other
places where we need to do similar cleanup.
2. Have main loop generate a signal and do the necessary cleanup as
handling that signal.

In order to keep the code clean and also let other places/future
changes leverage the same solution, approach #2 is used here to
define and generate a new signal TOOLS_CORE_SIG_NO_RPC. Also,
implement a handler for the same in the test plugin (for testing)
and dndcp for this bug.

While there also fixed the log domain for few files that are supposed
to logging under "dndcp" domain.