Tools hang detector now handles slow guest.
Tools hang detector can check its own past running history and figure
out whether itself is running slow due to a resource contention inside
the guest.

Tools hang detector shall report a different event to the VMX if the
main thread is not checking in and the tools hang detector itself is
not running frequently enough indicating the guest is running slow due
to a possible resource contention.