Common source file changes not directly applicable to open-vm-tools.
On Windows, the CDR tmp files under folder "VMwareDnD" will be kept
permanently in some cases after drag & drop.

Root cause: The existing strategy is to delete the folder before next reboot
of the machine, which is implemented through writing Windows HKLM registry.
However, for the non-administrator user, that user has no permission to
write the registry which in turn results in the temp files not being removed.

Solution: The temp files will be removed when the user disconnect the remote
desktop/app. The details are:
 1. Client will remove the temp folder when remote desktop/app is disconnected
    (rmks exits). Server will remove the temp folder when mksvchServer plugin
    gets "Not Ready" notification which means mksvchanServer is disconnected
    from the  mksvchanClient.
 2. Use prefix "Horizon_xxxx(pid)-" to distinguish if the temp folder is
    being used by DnD or not. For Client, the pid is the rmks process id,
    for Server, the pid is Clipboard pid.