Issue: Sometimes, there is a message "The system cannot find the file        specified" popped up when drag and drop over multiple remote apps.
Root cause:
When drag over multiple remote apps, multiple pairs of dragEnter and
dragLeave messages are sending. The DnD state or the dropSource/dataObject
are messed up when handling the 2nd DragEnter with the 1st DragLeave at
the same time.

Solution:
Besides fixing this issue, fixed several other issues also to enhance the
handling for multiple pairs of messages in the same time:
 1. In DnDController layer, avoid to reset the DnD state when handling the
    message responded from agent for previous sessions.
 2. In agent, only set the DnD state to be Ready when the previous
    DoDragDrop is really cancelled by OLE.
 3. In agent, only response to the button event when the drop is notified.
 4. In agent, add a 2s timeout checking for cancelling process to avoid
    conflicting with another DnDThread to create dropSrouce.
 5. In agent, add a 2s timeout checking for dragBegin process to avoid
    conflicting with previous dragBegin processing.
 6. Add virtual prefix to the method "GetData".