VIGOR and RPCI definitions for the tools hang detector events.
Define an array of ToolsHealthEvent to record the last N tools hang
events. By keeping a list of historical events since the power on,
we help ourselves trouble shooting guest/toolsd issues.

Use the first array element for the latest event. This simplifies
implementation as the DynArray can be easily capped by setting the
array count. This requires us to add a PushFront function to add
new element to the front of the DynArray.

Added the RPCI handler for the tools hang detector RPCI messages.

Added feature switch Tools_Update_Health.