GuestMapper: Detailed data fixes
The guestInfo detailed data for Photon was being reported incorrectly,
sometimes adding trailing whitespace when not needed.

The problem was how the release file was processed. It was being
processed (open/read/code/close) multiple times, and wasn't separating
each of the fields as it should. Fixed this.

Adding logging of what is sent by the guest; how things are mapped.