Don't send a backup manifest when aborting a Linux quiesced snapshot.
When taking a Linux quiesced snapshot, communication failures between
VMX and VMTools may result in VMTools sending a genericManifest event
message after the quiesced snapshot operation has been aborted.  If
this happens, VMX will send an error back to VMTools, which in turn
causes VMTools not to send genericManifest messages on subsequent
quiesced snapshots even if the host supports such messages.

One aspect of the implementation that gives rise to this behavior is
the use of the sync provider's snapshotDone function to undo a
quiescing operation.  Specifically, if VMTools aborts a quiesced
snapshot when the file system is quiesced, the quiescing must be
undone.  Currently, this is handled by calling the sync provider's
snapshotDone function.  This is the same function that is called to
complete the quiescing snapshot protocol when it is successful.  In
some respects this makes sense, since in either case snapshotDone
unquiesces the file system.  However, architecturally and conceptually,
it seems useful to distinguish between the action to be taken in the
successful case versus the aborting case.  It's also useful to do so
in practice, because the successful case sends the genericManifest
event to notify the host there is a backup manifest file, while the
aborting case should not do that.

To address the issue, add an "undo" function for the Linux sync
provider.  The undo function is called instead of snapshotDone as
part of aborting a quiesced snapshot in which the file system is
quiesced at the time of the abort.