Fix possible security issue with the permissions of the intermediate staging directory and path
/tmp/VMwareDnD is a staging directory used for DnD and CnP.  It should be
a regular directory, but malicious code or user may create the /tmp/VMwareDnD
as a symbolic link which points elsewhere on the system.  This may provide
user access to user B's files.

Do not set the permission of the root directory if the root directory
already exists and has the wrong permission.  The permission of the directory
must be 1777 if it is created by the VMToolsi.  If not, then the directory
has been created or modified by malicious code or user, so just cancel the
host to guest DnD or CnP operation.