Include vmware/tools/log.h to define g_info.
A recent change added a call to g_info from syncManifest.c.  This
in turn is causing open-vm-tools builds to fail because g_info is
not available on SLES 12sp1.

To fix the problem, include vmware/tools/log.h in syncManifest.c.
log.h defines g_info as a macro.