[Part 1] GuestLib support for 64bit memory shares.
MEMSCHED_SHARES_MAX is currently set to MAX_PPN and will not fit in 32-bit soon.

Changed the datatype of 'memshares' property to uint64 for the future support.

'memshares' is exposed in GuestLib as unsigned 32-bit value.
Modified the ESX code to set this property only if the vaulue can be fit in
a uint32 type. Else, the value is not set and 'valid' bit set is set to FALSE.

Added a new property "memShares64" (as uint64) in GuestLib for the new
64-bit memory shares.

This CLN implements only the necessary changes on ESX side.