[Part 2] GuestLib support for 64bit memory shares.
MEMSCHED_SHARES_MAX is currently set to MAX_PPN and will not fit in
32-bit soon. 'memshares' property which is currently available in GuestSDK
is an unsigned 32bit value.  Starting from ESX 7.0, this value is available
only if it can be fit in an unsigned 32-bit data type.

'memShares64' is the new unsigned 64-bit property added in GuestSDK for the
memory shares.

Added necessary APIs to retrieve the memshares64 value.  The new API
VMGuestLib_GetMemShares64 will first try to retrieve the memShares64 property.
If the underlying host is an older one, then the API will return the memShares
property.

Updated the vmGuestLibTest code to retrieve and print the new memShares64 value.