Remove the /etc/security directory from the guest vm-support bundle.
This directory contains sensitive guest side data that we should not
expose to the host side.

Although running the vm-support requires the root privilege in the guest,
the guest admin might not be aware of the ramification of running the script.
Therefore, better be cautious.

No file from the /etc/security has been needed in the past.  If any file is
needed in the future, that need will be evaluated case-by-case at that time.