DnD Phase II] - Avoid potential memory overflow.
In the case when dragging a format whose size is exactly the same as the
size threshold and plain text is provided, the plain text can't be added.
Current logic does not handle this case.  Instead it will result in a
large unsigned number (0-1 = 0xffff_ffff) of bytes to be allocated to
store plain text which causes memory overflow.

Just return in the case that no more size left to add plain text or the
plain text is empty.