Fix the 'procCmdName' attribute of process information.
* When listing down the processes, /proc/{PID}/cmdline file is read
and parsed to figure out the command name.  While doing this parsing,
the terminating NUL character is not parsed.  Due to this, if any
process doesn't have any command line arguments, the 'command name' is
retrieved as NULL.

Note: This issue doesn't happen if the cmdLine has any arguments.

Ex:
/usr/bin/vmtoolsd -> 'cmdName' is retrieved as NULL
/usr/bin/vmtoolsd -n vmusr -> 'cmdName' is retrieved properly as 'vmtoolsd'

* Fixed the code to include the trailing NUL character also while parsing.